Update migrations to use floats for necessary fields

+++ db/migrate/20160315191034_create_products.rb
@@ -3,7 +3,7 @@ class CreateProducts < ActiveRecord::Migration
     create_table :products do |t|
       t.text :name, null: false
       t.text :category
-      t.integer :serving_size, null: false
+      t.float :serving_size, null: false
       t.string :serving_unit, null: false
       t.integer :ndb_no, null: false, unique: true, index: true
